Add files via upload
This repository contains the source code to optimize and improve the performance of an image classification model on the CIFAR-10 dataset. The implementation is based on Transfer Learning techniques using DenseNet121, a pre-trained deep learning model. The goal is to leverage prior knowledge learned by DenseNet121 in other domains to improve efficiency and accuracy on the specific image classification task in CIFAR-10. It explores model tuning and customization strategies, as well as compilation and training techniques to achieve optimized results. This code is part of a project that seeks to advance image classification capabilities using knowledge transfer approaches.